requirements.txt
# Hugging Face / Transformers (Voxtral support currently on main branch)
git+https://github.com/huggingface/transformers@main
accelerate>=0.33.0
huggingface_hub>=0.23.4

# Audio/IO
soundfile>=0.12.1
librosa>=0.10.2.post1
numpy>=1.26.4
pydub>=0.25.1
requests>=2.32.0

# Diarization
pyannote.audio>=3.1.1

# Serverless runtime
runpod==1.6.2
